requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
Brotli==1.1.0
cachetools==5.5.0
cashfree_sdk==0.1.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==5.1.1
django-cors-headers==4.4.0
django-recaptcha==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.54.1
google==3.0.0
google-auth==2.35.0
html5lib==1.1
idna==3.10
lxml==5.3.0
oauthlib==3.2.2
oscrypto==1.3.0
pem==23.1.0
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pypdf==5.0.0
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
stripe==10.12.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3